+++ CHANGELOG.md
@@ -11,6 +11,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Add logging of 1-based `epoch` to the console and the loggers.
 - Add `MixUp` augmentation for LT-DETR object detection training.
+- Add support for COCO object detection dataset format.
 
 ### Changed